+++ apigateway/src/main/java/com/node5/apigateway/config/ReactiveAuthorization.java
@@ -12,12 +12,10 @@ import org.springframework.security.authorization.ReactiveAuthorizationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.server.authorization.AuthorizationContext;
 import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
-import java.util.UUID;
 
 
 @Slf4j
@@ -25,11 +23,9 @@ import java.util.UUID;
 public class ReactiveAuthorization implements ReactiveAuthorizationManager<AuthorizationContext> {
 
     private final SecretKey secretKey;
-    private final WebClient.Builder webClientBuilder;
 
-    public ReactiveAuthorization(@Value("${token.secret}") String secret, WebClient.Builder builder) {
+    public ReactiveAuthorization(@Value("${token.secret}") String secret) {
         this.secretKey = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
-        this.webClientBuilder = builder;
     }
 
     @Override
@@ -65,31 +61,9 @@ public class ReactiveAuthorization implements ReactiveAuthorizationManager<Autho
 
             context.getExchange().getAttributes().put("cached_claims", claims);
 
-            return authorizeByMemberService(claims.getSubject(), request)
-                    .map(AuthorizationDecision::new)
-                    .onErrorReturn(new AuthorizationDecision(false));
+            return Mono.just(new AuthorizationDecision(true));
         } catch (Exception e) {
             return Mono.just(new AuthorizationDecision(false));
         }
-    }
-
-    private Mono<Boolean> authorizeByMemberService(String memberId, ServerHttpRequest request) {
-        return webClientBuilder.build()
-                .post()
-                .uri("lb://member-service/internal/auth/authorize")
-                .bodyValue(new AuthorizeRequest(
-                        UUID.fromString(memberId),
-                        request.getMethod().name(),
-                        request.getPath().pathWithinApplication().value()
-                ))
-                .retrieve()
-                .bodyToMono(Boolean.class);
-    }
-
-    record AuthorizeRequest(
-            UUID memberId,
-            String method,
-            String path
-    ) {
     }
 }
