+++ apigateway/src/main/java/com/node5/apigateway/filter/AuthenticationFilter.java
@@ -7,10 +7,6 @@ import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFac
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.stereotype.Component;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
 @Component
 public class AuthenticationFilter extends AbstractGatewayFilterFactory<AuthenticationFilter.Config> {
     public AuthenticationFilter() {
@@ -19,23 +15,23 @@ public class AuthenticationFilter extends AbstractGatewayFilterFactory<Authentic
 
     @Override
     public GatewayFilter apply(AuthenticationFilter.Config config) {
-        return ((exchange, chain) -> {
+        return (exchange, chain) -> {
             Object attribute = exchange.getAttribute("cached_claims");
 
-            if(attribute instanceof Claims claims) {
-                String memberId = claims.getSubject();
-                String status = claims.get("memberStatus", String.class);
-
-                ServerHttpRequest request = exchange.getRequest().mutate()
-                        .header("Member-Id", memberId)
-                        .header("Member-Status", status)
-                        .build();
-
-                return chain.filter(exchange.mutate().request(request).build());
+            if (!(attribute instanceof Claims claims)) {
+                return chain.filter(exchange);
             }
 
-            return chain.filter(exchange);
-        });
+            String memberId = claims.getSubject();
+            String status = claims.get("memberStatus", String.class);
+
+            ServerHttpRequest request = exchange.getRequest().mutate()
+                    .header("Member-Id", memberId)
+                    .header("Member-Status", status)
+                    .build();
+
+            return chain.filter(exchange.mutate().request(request).build());
+        };
     }
 
     @Data
