+++ apigateway/src/main/resources/application-prod.yaml
@@ -55,6 +55,7 @@ spring:
                 preLogger: true
                 postLogger: true
             - name: AuthenticationFilter
+            - name: AuthorizationFilter
           discovery:
             locator:
               enabled: true
